+++ lib/pages/mappage/mappage.dart
@@ -145,6 +145,7 @@ class _MapPageState extends State<MapPage> {
               },
             ),
           ),
+
         ],
       ),
     );
